Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/directive-parsing-04-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src:'.

A warning should be logged to the console, as `script-src: 'none'` shouldn't contain a colon. Since the directive is invalid, the script should run.



--------
Frame: '<!--frame1-->'
--------
PASS
